+++ src/test/java/ru/netology/stats/statis/service/StatsServiceTest.java
@@ -9,16 +9,16 @@ public class StatsServiceTest {
     @Test
     public void numberMonthMinSales() {
         StatsService service = new StatsService();
-        int[] sales = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedMonthMinSales = 9;
-        int actualMonthMinSales = service.minSales(sales);
+        long actualMonthMinSales = service.minSales(sales);
         Assertions.assertEquals(expectedMonthMinSales, actualMonthMinSales);
     }
 
     @Test
     public void sumOfAllSales() {
         StatsService service = new StatsService();
-        int[] sales = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         long expectedSumSales = 180;
         long actualSumSales = service.sumAllSales(sales);
         Assertions.assertEquals(expectedSumSales, actualSumSales);
@@ -27,7 +27,7 @@ public class StatsServiceTest {
     @Test
     public void averageSales() {
         StatsService service = new StatsService();
-        int[] sales = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         long expectedAverageSales = 15;
         long actualAverageSales = service.averageSummSales(sales);
         Assertions.assertEquals(expectedAverageSales, actualAverageSales);
@@ -36,16 +36,16 @@ public class StatsServiceTest {
     @Test
     public void numberLastMonthMaxSales() {
         StatsService service = new StatsService();
-        int[] sales = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedMonthMaxSales = 8;
-        int actualMonthMaxSales = service.maxSales(sales);
+        long actualMonthMaxSales = service.maxSales(sales);
         Assertions.assertEquals(expectedMonthMaxSales, actualMonthMaxSales);
     }
 
     @Test
     public void NumbMonthsBelowAverage() {
         StatsService service = new StatsService();
-        int[] sales = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedBelowAverage = 5;
         int actualBelowAverage = service.monthsBelowAverage(sales);
         Assertions.assertEquals(expectedBelowAverage, actualBelowAverage);
@@ -54,7 +54,7 @@ public class StatsServiceTest {
     @Test
     public void NumbMonthsAboveAverage() {
         StatsService service = new StatsService();
-        int[] sales = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedAboveAverage = 5;
         int actualAboveAverage = service.monthsAboveAverage(sales);
         Assertions.assertEquals(expectedAboveAverage, actualAboveAverage);
